+++ src/app/imports/store/reducers/index.js
@@ -9,7 +9,7 @@ import { matchesByMatchup } from './_matchupMatches';
 import { matchesByParticipant } from './_participantMatches';
 import { participantsHeatmapByParticipant } from './_participantsHeatmap';
 import { participantsPerformanceByParticipant } from './_participantsPerformance';
-import { playedTogetherBySummonerIds } from './_playedTogether';
+import { playedTogetherBySummonerNames } from './_playedTogether';
 import { statsByMatchup } from './_matchupStats';
 import { trophyStatsByTrophyName } from './_trophyStats';
 import { uiStates } from './_uiStates';
@@ -26,7 +26,7 @@ const reducers = combineReducers({
   matchesByParticipant,
   participantsHeatmapByParticipant,
   participantsPerformanceByParticipant,
-  playedTogetherBySummonerIds,
+  playedTogetherBySummonerNames,
   statsByMatchup,
   uiStates
 });